# docs/api/openapi/mesas.openapi.yaml
...
openapi: "3.0.3"
info:
  title: "Mesas API — Artifício RPG"
  description: "API do módulo Mesas. Metadados x-artificio-* definidos por heurística de path."
  version: "0.1.0"
servers:
  - url: https://mesas.artificiorpg.com
    description: Produção
  - url: http://localhost:4000
    description: Desenvolvimento
security: []
paths:
...
  /api/v1/admin/discord/fetch:
    post:
      operationId: postapi_v1_admin_discord_fetch
      summary: "Cria ou executa api v1 admin discord fetch"
      x-artificio-owner: mesas
      x-artificio-scope: admin
      x-artificio-status: active
      x-artificio-auth: admin
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...
  /api/v1/admin/discord/sources/{sourceId}/reingest-force:
    post:
      operationId: postapi_v1_admin_discord_sources_sourceId_reingest_force
      summary: "Cria ou executa api v1 admin discord sources sourceId reingest-force"
      x-artificio-owner: mesas
      x-artificio-scope: admin
      x-artificio-status: active
      x-artificio-auth: admin
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...